Updated Classes Cascade Mapping and Bootstrap
For now, until further notice from Babcock, I updated the cascade mapping so that when you delete a class it does not delete the indicators. Only when an outcome is delete does the assoicated indicators get deleted as well.

Also, I changed the bootstrap a little so that now there will be a default outcome and 2 indicators linked to it, which only happens if outcomes.count() is 0 which means that there is no currently existing data in the DB