A single test case is now running for AD controller!!!
There is now one test case that runs for the AD controller that tests if a request with correct params and "post" method are set and the index() method is ran on AD controller, than a new AD should be created and saved and retrievable, which this test case shows. More test cases are to come. This test case was to help me learn how to write a controller test case with grails properly.