Refactoring of Classes to Courses is building
At this current stage of the refactoring I have Renamed the Classes domain to Courses, and I have done enough refactoring that the project is able to build and run on localHost. The rest of the refactoring for the other controllers and pages still need to be done and tested.