Further progress towards refactoring Classes to Courses
At this stage I have refactored all of the neccessary code for the project to build as well as for the project to function correctly.  This entails changing all of the references to the old "Classes" domain to reference the new "Courses" domain. Essentially all of the code is referencing correctly, the only thing left to complete this refactoring process is to change all of the variable names to better represent "course" or "courses" vs "class" or "classes". Other than that, functionality is back on track with the refactoring process.